+++ src/frequenz/sdk/timeseries/__init__.py
@@ -37,10 +37,12 @@ Example:
 
 from ._base_types import UNIX_EPOCH, Sample, Sample3Phase
 from ._moving_window import MovingWindow
+from ._periodic_feature_extractor import PeriodicFeatureExtractor
 from ._resampling import ResamplerConfig
 
 __all__ = [
     "MovingWindow",
+    "PeriodicFeatureExtractor",
     "ResamplerConfig",
     "Sample",
     "Sample3Phase",