+++ src/frequenz/sdk/timeseries/_moving_window.py
@@ -164,7 +164,8 @@ class MovingWindow:
             input_sampling_period <= size
         ), "The input sampling period should be equal to or lower than the window size."
 
-        sampling = input_sampling_period
+        self._sampling_period = input_sampling_period
+
         self._resampler: Resampler | None = None
         self._resampler_sender: Sender[Sample] | None = None
         self._resampler_task: asyncio.Task[None] | None = None
@@ -175,15 +176,17 @@ class MovingWindow:
             ), "The resampling period should be equal to or lower than the window size."
 
             self._resampler = Resampler(resampler_config)
-            sampling = resampler_config.resampling_period
+            self._sampling_period = resampler_config.resampling_period
 
         # Sampling period might not fit perfectly into the window size.
-        num_samples = math.ceil(size.total_seconds() / sampling.total_seconds())
+        num_samples = math.ceil(
+            size.total_seconds() / self._sampling_period.total_seconds()
+        )
 
         self._resampled_data_recv = resampled_data_recv
         self._buffer = OrderedRingBuffer(
             np.empty(shape=num_samples, dtype=float),
-            sampling_period=sampling,
+            sampling_period=self._sampling_period,
             align_to=align_to,
         )
 
@@ -193,6 +196,16 @@ class MovingWindow:
         self._update_window_task: asyncio.Task[None] = asyncio.create_task(
             self._run_impl()
         )
+
+    @property
+    def sampling_period(self) -> timedelta:
+        """
+        Return the sampling period of the MovingWindow.
+
+        Returns:
+            The sampling period of the MovingWindow.
+        """
+        return self._sampling_period
 
     async def _run_impl(self) -> None:
         """Awaits samples from the receiver and updates the underlying ring buffer.
@@ -292,7 +305,17 @@ class MovingWindow:
             an numpy array if the key is a slice.
         """
         if isinstance(key, slice):
+            if isinstance(key.start, int) or isinstance(key.stop, int):
+                if key.start is None or key.stop is None:
+                    key = slice(slice(key.start, key.stop).indices(self.__len__()))
+            elif isinstance(key.start, datetime) or isinstance(key.stop, datetime):
+                if key.start is None:
+                    key = slice(self._buffer.time_bound_oldest, key.stop)
+                if key.stop is None:
+                    key = slice(key.start, self._buffer.time_bound_newest)
+
             _logger.debug("Returning slice for [%s:%s].", key.start, key.stop)
+
             # we are doing runtime typechecks since there is no abstract slice type yet
             # see also (https://peps.python.org/pep-0696)
             if isinstance(key.start, datetime) and isinstance(key.stop, datetime):